factor out one off page templates to top level pages and components
