+++ src/pages/about.html
@@ -1,0 +1,14 @@
+<!DOCTYPE html>
+<html lang="en" prefix="og:http://ogp.me/ns#">
+
+  <head>
+    <script type="module" src="../components/content-list-component/content-list-component.js"></script>
+  </head>
+
+  <body>
+    <p>I think the best way to tell you about myself is to show you what I am passionate about. Below are some featured articles and presentations I've worked on.</p>
+    
+    <app-content-list></app-content-list>
+  </body>
+
+</html>
