--- src/pages/about.md
@@ -1,5 +1,0 @@
----
-template: 'about'
----
-
-I think the best way to tell you about myself is to show you what I am passionate about. Below are some featured articles and presentations I've worked on.
