+++ src/pages/projects.html
@@ -1,0 +1,14 @@
+<!DOCTYPE html>
+<html lang="en" prefix="og:http://ogp.me/ns#">
+
+  <head>
+    <script type="module" src="../components/projects-list-component/projects-list-component.js"></script>
+  </head>
+
+  <body>
+    <p>These are a few projects that I think reflect some of my general interests and activities.</p>
+    
+    <app-projects-list></app-projects-list>
+  </body>
+
+</html>
