--- src/pages/projects.md
@@ -1,5 +1,0 @@
----
-template: 'projects'
----
-
-These are a few projects that I think reflect some of my general interests and activities.
