--- src/templates/projects.html
@@ -1,16 +1,0 @@
-<!DOCTYPE html>
-<html lang="en" prefix="og:http://ogp.me/ns#">
-
-  <head>
-    <script type="module" src="/templates/projects.js"></script>
-  </head>
-
-  <body>
-    <!-- TODO: could move markdown content into here -->
-    <!-- or could move <app-projects-list> to components and / or markdown using frontmatter import -->
-    <!-- or when nested paths work -->
-    <content-outlet></content-outlet>
-    <app-projects-list></app-projects-list>
-  </body>
-
-</html>
